+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - Fast and dirty mode, to find quickly any path
+- May check new obstacles while searching or not
 
 # v1.4.1
 - May use an external physics engine